# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Medicine Reminder REST API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
}

export type MedicineFrequency = typeof MedicineFrequency[keyof typeof MedicineFrequency];


export const MedicineFrequency = {
  once_daily: 'once_daily',
  twice_daily: 'twice_daily',
  three_daily: 'three_daily',
  custom: 'custom',
} as const;

export interface Medicine {
  id: number;
  medicineName: string;
  /** @nullable */
  tabletImage?: string | null;
  dosage: string;
  frequency: MedicineFrequency;
  reminderTimes: string[];
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
}

export type MedicineInputFrequency = typeof MedicineInputFrequency[keyof typeof MedicineInputFrequency];


export const MedicineInputFrequency = {
  once_daily: 'once_daily',
  twice_daily: 'twice_daily',
  three_daily: 'three_daily',
  custom: 'custom',
} as const;

export interface MedicineInput {
  /** @minLength 1 */
  medicineName: string;
  /** @nullable */
  tabletImage?: string | null;
  /** @minLength 1 */
  dosage: string;
  frequency: MedicineInputFrequency;
  /**
     * @minItems 1
     * @items.pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$
     */
  reminderTimes: string[];
  startDate: string;
  /** @nullable */
  endDate?: string | null;
}

export type MedicineUpdateFrequency = typeof MedicineUpdateFrequency[keyof typeof MedicineUpdateFrequency];


export const MedicineUpdateFrequency = {
  once_daily: 'once_daily',
  twice_daily: 'twice_daily',
  three_daily: 'three_daily',
  custom: 'custom',
} as const;

export interface MedicineUpdate {
  /** @minLength 1 */
  medicineName?: string;
  /** @nullable */
  tabletImage?: string | null;
  /** @minLength 1 */
  dosage?: string;
  frequency?: MedicineUpdateFrequency;
  /**
     * @minItems 1
     * @items.pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$
     */
  reminderTimes?: string[];
  startDate?: string;
  /** @nullable */
  endDate?: string | null;
}

export type ReminderStatus = typeof ReminderStatus[keyof typeof ReminderStatus];


export const ReminderStatus = {
  upcoming: 'upcoming',
  due: 'due',
  taken: 'taken',
  skipped: 'skipped',
} as const;

export interface Reminder {
  id: number;
  medicineId: number;
  medicineName: string;
  /** @nullable */
  tabletImage?: string | null;
  dosage: string;
  reminderTime: string;
  reminderDate: string;
  status: ReminderStatus;
  /** @nullable */
  takenAt?: string | null;
}

export interface ReminderInput {
  medicineId: number;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  reminderTime: string;
  reminderDate: string;
}

export interface DashboardSummary {
  today: Reminder[];
  nextReminder: Reminder | null;
  takenCount: number;
  totalCount: number;
  adherencePercent: number;
}

export interface TabletIdentifyInput {
  /** @minLength 1 */
  imageData: string;
}

export interface TabletIdentification {
  label: string;
  confidence: number;
  isDemo: boolean;
  disclaimer: string;
}

export type ListRemindersParams = {
filter?: ListRemindersFilter;
};

export type ListRemindersFilter = typeof ListRemindersFilter[keyof typeof ListRemindersFilter];


export const ListRemindersFilter = {
  today: 'today',
  upcoming: 'upcoming',
  completed: 'completed',
  skipped: 'skipped',
  all: 'all',
} as const;