# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Medicine Reminder REST API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Session authentication
  - name: dashboard
    description: Dashboard summaries
  - name: medicines
    description: Medicine management
  - name: reminders
    description: Scheduled dose tracking
  - name: tablet
    description: Clearly labeled prototype tablet identification
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      operationId: login
      tags: [auth]
      summary: Start a user session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      responses:
        "204":
          description: Logged out
  /me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get today's medicine and reminder summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /medicines:
    get:
      operationId: listMedicines
      tags: [medicines]
      summary: List the signed-in user's medicines
      responses:
        "200":
          description: Medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
    post:
      operationId: createMedicine
      tags: [medicines]
      summary: Add a medicine and create its reminders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "201":
          description: Medicine created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /medicines/{id}:
    get:
      operationId: getMedicine
      tags: [medicines]
      summary: Get one medicine
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
        "404":
          description: Medicine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMedicine
      tags: [medicines]
      summary: Update a medicine
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineUpdate"
      responses:
        "200":
          description: Updated medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
        "404":
          description: Medicine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMedicine
      tags: [medicines]
      summary: Delete a medicine and its reminders
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "404":
          description: Medicine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List reminders for the signed-in user
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [today, upcoming, completed, skipped, all]
            default: all
      responses:
        "200":
          description: Reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder for a medicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}/taken:
    patch:
      operationId: markReminderTaken
      tags: [reminders]
      summary: Mark a reminder as taken
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}/skipped:
    patch:
      operationId: markReminderSkipped
      tags: [reminders]
      summary: Mark a reminder as skipped
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}/dismiss:
    patch:
      operationId: dismissReminder
      tags: [reminders]
      summary: Dismiss the active browser alarm without changing status
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /tablet/identify:
    post:
      operationId: identifyTablet
      tags: [tablet]
      summary: Run the demo-only tablet identification flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TabletIdentifyInput"
      responses:
        "200":
          description: Demo identification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TabletIdentification"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
      required: [id, name, email]
    RegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [name, email, password]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
    Medicine:
      type: object
      properties:
        id:
          type: integer
        medicineName:
          type: string
        tabletImage:
          type: ["string", "null"]
        dosage:
          type: string
        frequency:
          type: string
          enum: [once_daily, twice_daily, three_daily, custom]
        reminderTimes:
          type: array
          items:
            type: string
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        [id, medicineName, dosage, frequency, reminderTimes, startDate, createdAt]
    MedicineInput:
      type: object
      properties:
        medicineName:
          type: string
          minLength: 1
        tabletImage:
          type: ["string", "null"]
        dosage:
          type: string
          minLength: 1
        frequency:
          type: string
          enum: [once_daily, twice_daily, three_daily, custom]
        reminderTimes:
          type: array
          minItems: 1
          items:
            type: string
            pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
      required:
        [medicineName, dosage, frequency, reminderTimes, startDate]
    MedicineUpdate:
      type: object
      properties:
        medicineName:
          type: string
          minLength: 1
        tabletImage:
          type: ["string", "null"]
        dosage:
          type: string
          minLength: 1
        frequency:
          type: string
          enum: [once_daily, twice_daily, three_daily, custom]
        reminderTimes:
          type: array
          minItems: 1
          items:
            type: string
            pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
    Reminder:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        tabletImage:
          type: ["string", "null"]
        dosage:
          type: string
        reminderTime:
          type: string
        reminderDate:
          type: string
          format: date
        status:
          type: string
          enum: [upcoming, due, taken, skipped]
        takenAt:
          type: ["string", "null"]
          format: date-time
      required:
        [id, medicineId, medicineName, dosage, reminderTime, reminderDate, status]
    ReminderInput:
      type: object
      properties:
        medicineId:
          type: integer
        reminderTime:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        reminderDate:
          type: string
          format: date
      required: [medicineId, reminderTime, reminderDate]
    DashboardSummary:
      type: object
      properties:
        today:
          type: array
          items:
            $ref: "#/components/schemas/Reminder"
        nextReminder:
          anyOf:
            - $ref: "#/components/schemas/Reminder"
            - type: "null"
        takenCount:
          type: integer
        totalCount:
          type: integer
        adherencePercent:
          type: integer
      required: [today, nextReminder, takenCount, totalCount, adherencePercent]
    TabletIdentifyInput:
      type: object
      properties:
        imageData:
          type: string
          minLength: 1
      required: [imageData]
    TabletIdentification:
      type: object
      properties:
        label:
          type: string
        confidence:
          type: integer
        isDemo:
          type: boolean
        disclaimer:
          type: string
      required: [label, confidence, isDemo, disclaimer]