# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Medicine Reminder REST API
 * OpenAPI spec version: 0.1.0
 */
import type { Reminder } from './reminder';

export interface DashboardSummary {
  today: Reminder[];
  nextReminder: Reminder | null;
  takenCount: number;
  totalCount: number;
  adherencePercent: number;
}