# lib/api-zod/src/generated/types/medicine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Medicine Reminder REST API
 * OpenAPI spec version: 0.1.0
 */
import type { MedicineFrequency } from './medicineFrequency';

export interface Medicine {
  id: number;
  medicineName: string;
  /** @nullable */
  tabletImage?: string | null;
  dosage: string;
  frequency: MedicineFrequency;
  reminderTimes: string[];
  startDate: Date;
  /** @nullable */
  endDate?: Date | null;
  createdAt: Date;
}