# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Smart Medicine Reminder REST API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}